Change rezulte

+++ src/Main.java
@@ -4,7 +4,7 @@ public class Main {
         BmiService service = new BmiService();
         double height = 1.87;
         int weight = 98;
-        double bmi = service.calculate(height, weight); //должно получиться 28
+        int bmi = service.calculate(height, weight); //должно получиться 28
         System.out.println();
         System.out.println("Индекс массы тела:");
         System.out.println(bmi);
@@ -13,3 +13,5 @@ public class Main {
         System.out.println(service.calculate(1.53,51));
     }
 }
+
+
